Remove GovGrOAuthOptions and only use GovGrOptions for configuration
