+++ src/Indice.AspNetCore.Authentication.GovGr/GovGrHandler.cs
@@ -1,10 +1,15 @@
+using System.Security.Claims;
 using System.Text.Encodings.Web;
+using System.Text.Json;
+using System.Xml.Linq;
+using Indice.Security;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.OAuth;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Polly;
 
 namespace Indice.AspNetCore.Authentication.GovGr;
 
@@ -22,6 +27,35 @@ public class GovGrHandler : OAuthHandler<GovGrOptions>, IAuthenticationSignOutHa
         UrlEncoder encoder) 
         : base(options, logger, encoder)
     {
+    }
+
+    ///<summary>
+    ///Creates an authentication ticket for the user.
+    ///</summary>
+
+    protected override async Task<AuthenticationTicket> CreateTicketAsync(ClaimsIdentity identity, AuthenticationProperties properties, OAuthTokenResponse tokens) {
+
+        using (var user = JsonDocument.Parse("{}")) {
+            var context = new OAuthCreatingTicketContext(new ClaimsPrincipal(identity), properties, Context, Scheme, Options, Backchannel, tokens, user.RootElement);
+            var accessToken = context.Properties.GetTokenValue("access_token");
+            var httpClient = context.Backchannel;
+            using var request = new HttpRequestMessage(HttpMethod.Get, context.Options.UserInformationEndpoint);
+            request.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", accessToken);
+            var response = await httpClient.SendAsync(request);
+            response.EnsureSuccessStatusCode();
+            var responseBody = await response.Content.ReadAsStringAsync();
+            var xml = XDocument.Parse(responseBody);
+            var claims = xml.Descendants("userinfo")
+                            .SelectMany(x => x.Attributes()
+                                              .Select(attr => new Claim(GovGrExtensions.GovGrClaimMap.GetValueOrDefault(attr.Name.LocalName, attr.Name.LocalName), attr.Value.Trim())))
+                            .Where(x => !GovGrExtensions.GovGrClaimNullLiteral.Equals(x.Value, StringComparison.OrdinalIgnoreCase))
+                            .ToList();
+            // add another claim for subject since this is not available.
+            claims.Add(new Claim(BasicClaimTypes.Subject, claims.Find(x => x.Type == BasicClaimTypes.Name)!.Value));
+            context.Principal = new ClaimsPrincipal(new ClaimsIdentity(claims, context.Scheme.Name, BasicClaimTypes.Name, BasicClaimTypes.Role));
+            await Events.CreatingTicket(context);
+            return new AuthenticationTicket(context.Principal!, context.Properties, Scheme.Name);
+        }
     }
 
     /// <summary>
